Fix problem with missing image height and size
The plugin threw a lot of error messages in my code. $height/$width has not been initialized. Return 'false' for width and height, when not available.
For me the problem was, that the submitted size was not available. So it threw an error in line 149. Should be fixed now supposing, that returning false for width and height is catched/accepted from the calling function.